Add files via upload
The autoAssign function operates using two main queues:

workQueue: This contains all workers qualified to perform tasks on any given Sunday. Workers are prioritized in the queue, with the highest-priority worker at the front and the lowest-priority worker at the back. After a worker is assigned a task, they are moved to the end of the queue for the next Sunday. Each week, the workQueue is cross-referenced with the availableQueue to determine task assignments.

availableQueue: This represents the workers available on a specific Sunday, considering special constraints like husband-and-wife teams etc. It essentially serves as the pull-down list for the worship schedule.

The autoAssign function operates in two modes:

Round Robin Mode: This is the basic algorithm where workers are assigned tasks in a rotating order. Once a worker is assigned, they move to the end of the queue for the following week.

Balance Mode: This mode enhances the round robin method by introducing a third queue, workerCount, which tracks how many times each worker has been assigned a task during the current quarter. The priority in workQueue is then adjusted based on the number of assignments, with workers who have been assigned less frequently moved to the front of the queue.

Extensive testing shows that the balance mode outperforms the round robin mode, leading to a more even distribution of tasks across workers.